+++ include/ldl_radio_debug.h
@@ -19,40 +19,43 @@
  *
  * */
 
-#ifndef MBED_LDL_PORT_H
-#define MBED_LDL_PORT_H
-
-#include "mbed_assert.h"
-#include "mbed_trace.h"
-
-#define TRACE_GROUP "LDL"
-
-#define LDL_ASSERT(X) MBED_ASSERT(X);
-
-#define LDL_DEBUG(APP, ...) do{\
-    tr_debug(__VA_ARGS__);\
-}while(0);
-
-#define LDL_ERROR(APP, ...) do{\
-    tr_error( __VA_ARGS__);\
-}while(0);
+#ifndef LDL_RADIO_DEBUG_H
+#define LDL_RADIO_DEBUG_H
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-void my_trace_begin(void);
-void my_trace_part(const char *fmt, ...);
-void my_trace_hex(const uint8_t *ptr, size_t len);
-void my_trace_end(void);
+#include <stdint.h>
+#include <stdbool.h>
+#include <stddef.h>
+
+#ifndef LDL_RADIO_DEBUG_LOG_SIZE
+#define LDL_RADIO_DEBUG_LOG_SIZE 40
+#endif
+
+struct ldl_radio;
+
+struct ldl_radio_debug_log_entry {
+
+    uint8_t opcode;
+    uint8_t value;
+    uint8_t size;
+};
+
+struct ldl_radio_debug_log {
+
+    struct ldl_radio_debug_log_entry log[LDL_RADIO_DEBUG_LOG_SIZE];
+    uint8_t pos;
+    bool overflow;
+};
+
+void LDL_Radio_debugLogReset(struct ldl_radio *self);
+void LDL_Radio_debugLogPush(struct ldl_radio *self, uint8_t opcode, const uint8_t *data, uint8_t size);
+void LDL_Radio_debugLogFlush(struct ldl_radio *self, const char *fn);
 
 #ifdef __cplusplus
 }
 #endif
-
-#define LDL_TRACE_BEGIN() my_trace_begin();
-#define LDL_TRACE_PART(...) my_trace_part(__VA_ARGS__);
-#define LDL_TRACE_HEX(PTR, LEN) my_trace_hex(PTR, LEN);
-#define LDL_TRACE_FINAL() my_trace_end();
 
 #endif